interface 추가, 카테고리 인터페이스화
