+++ CHATGAME/Assets/Scripts/Game/AffectionPat.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8b8fbdb0afcbdd14d8f2e3fc446345b7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
