+++ CHATGAME/Assets/Scripts/Game/AffectionTwt.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 95a5660b9b9e25f47bd296272b9a25aa
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
