+++ CHATGAME/Assets/Scripts/Game/ICategory.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f4536c4fefa91ed43aba82693e0d9d0b
+guid: e2db29664a654b54b84940e6e28caa01
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
